requirements.txt
numpy==1.23.5
pandas==1.5.3
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.2.2
matplotlib==3.6.3
seaborn==0.12.2
plotly==5.13.0
tqdm==4.65.0
python-dotenv==1.0.0
pyyaml==6.0